requirements.txt
fastapi>=0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
beautifulsoup4==4.12.3
aiocache==0.12.2
python-dotenv==1.0.1
pytoniq==0.1.41
pydantic>=2.0.0
